Moved tests back into original directory to make this easier to review.
