+++ improver_tests/ensemble_copula_coupling/ensemble_copula_coupling/ecc_test_data.py
