+++ improver_tests/ensemble_copula_coupling/ensemble_copula_coupling/test_ConvertProbabilitiesToPercentiles.py
@@ -46,7 +46,7 @@ from improver.ensemble_copula_coupling.ensemble_copula_coupling import (
 from improver.metadata.probabilistic import find_threshold_coordinate
 from improver.utilities.warnings_handler import ManageWarnings
 
-from ..calibration.ensemble_calibration.helper_functions import (
+from ...calibration.ensemble_calibration.helper_functions import (
     add_forecast_reference_time_and_forecast_period,
     set_up_probability_above_threshold_spot_temperature_cube,
     set_up_probability_above_threshold_temperature_cube,
