+++ improver_tests/ensemble_copula_coupling/ensemble_copula_coupling/test_ResamplePercentiles.py
@@ -43,7 +43,7 @@ from improver.ensemble_copula_coupling.ensemble_copula_coupling import (
 )
 from improver.utilities.warnings_handler import ManageWarnings
 
-from ..set_up_test_cubes import (
+from ...set_up_test_cubes import (
     add_coordinate,
     set_up_percentile_cube,
     set_up_variable_cube,
