+++ improver_tests/ensemble_copula_coupling/ensemble_copula_coupling/test_utilities.py
@@ -51,7 +51,7 @@ from improver.ensemble_copula_coupling.utilities import (
     restore_non_probabilistic_dimensions,
 )
 
-from ..set_up_test_cubes import set_up_percentile_cube, set_up_variable_cube
+from ...set_up_test_cubes import set_up_percentile_cube, set_up_variable_cube
 from .ecc_test_data import ECC_TEMPERATURE_REALIZATIONS, set_up_spot_test_cube
